reorganized qcd_ml.base.paths
