+++ tox.ini
@@ -6,7 +6,7 @@ toxworkdir=/tmp/.tox
 [testenv]
 wheel = true
 setenv = 
-    OMP_NUM_THREADS=8
+    OMP_NUM_THREADS=6
     OMP_PROC_BIND=spread
     OMP_PLACES=cores
 deps =
